+++ src/components/ExportComponent/style.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-const StyledDiv = styled.div`
+export const StyledDiv = styled.div`
   margin: 0 0 2rem;
 `;
 
@@ -10,4 +10,14 @@ export const StyledSpan = styled.span`
   margin: 0.5rem 0;
 `;
 
-export default StyledDiv;
+export const StyledRelativeDiv = styled.div`
+  position: relative;
+`;
+
+export const StyledClipboardWrapper = styled.div`
+  display: inline-block;
+  position: absolute;
+  top: 50%;
+  transform: translateY(-50%);
+  margin: 0 0.5rem;
+`;