+++ CHANGELOG.md
@@ -4,16 +4,28 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+<br/>
+
 ## `Unreleased`
 
-- Add functionality to copy theme to clipboard.
 - Add reset functionality to revert changes.
 - Migrate state in Panel to Context.
+
+<br/>
+
+## `0.7.1` - 2020 July 27
+
+- Added functionality to copy changed theme to clipboard.
+- Added `CopyToClipboard` component.
+
+<br/>
 
 ## `0.7.0` - 2020 July 27
 
 - Added functionality to export changed theme files.
 - Added `Export` component.
+
+<br/>
 
 ## `0.6.0` - 2020 July 26
 
@@ -21,6 +33,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added `Button`, `CodePreview`, `PlaygroundStage`, `ReviewStage`,
 - Added wrapper around panel with ref to control scrolling.
 - Added `example` script to run the example directly.
+
+<br/>
 
 ## `0.5.0` - 2020 July 24
 
@@ -31,11 +45,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Changed `ThemingPanel` component to `Panel`.
 - Removed `rimraf` package.
 
+<br/>
+
 ## `0.4.0` - 2020 July 23
 
 - Panel generates input fields based on the theme files obtained.
 - Added `FormGroup`, `Heading`, `Label`, `SelectInput`, `TextInput`, `ThemingFields` components.
 - Installed `eslint-plugin-react-hooks` plugin.
+
+<br/>
 
 ## `0.3.1` - 2020 July 21
 
@@ -43,6 +61,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added `ThemingPanel` component.
 - Added `REQUEST_INITIAL_THEME` and `INITIAL_THEME` events in `events.js`.
 - Removed `prepare` from scripts.
+
+<br/>
 
 ## `0.3.0` - 2020 July 21
 
@@ -52,10 +72,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Installed `styled-components`.
 - Installed `babel-plugin-styled-components` plugin.
 
+<br/>
+
 ## `0.2.0` - 2020 July 20
 
 - Added Theming panel in Storybook.
 - Added `.nvmrc` with Node version.
+
+<br/>
 
 ## `0.1.0` - 2020 July 19